Move cmShell into a dir
